+++ ziti/ziti_test.go
@@ -32,10 +32,11 @@ func Test_contextImpl_processServiceUpdates(t *testing.T) {
 		options: &Options{
 			OnServiceUpdate: servUpdate,
 		},
-		services:     cmap.New[*rest_model.ServiceDetail](),
-		sessions:     cmap.New[*rest_model.SessionDetail](),
-		intercepts:   cmap.New[*edge.InterceptV1Config](),
-		EventEmmiter: events.New(),
+		services:           cmap.New[*rest_model.ServiceDetail](),
+		sessions:           cmap.New[*rest_model.SessionDetail](),
+		serviceEdgeRouters: cmap.New[[]*rest_model.SessionEdgeRouter](),
+		intercepts:         cmap.New[*edge.InterceptV1Config](),
+		EventEmmiter:       events.New(),
 	}
 
 	ctx.CtrlClt = &CtrlClient{
@@ -220,10 +221,11 @@ func Test_AddressMatch(t *testing.T) {
 
 	ctx := &ContextImpl{
 		options:      &Options{},
-		services:     cmap.New[*rest_model.ServiceDetail](),
-		sessions:     cmap.New[*rest_model.SessionDetail](),
-		intercepts:   cmap.New[*edge.InterceptV1Config](),
-		EventEmmiter: events.New(),
+		services:           cmap.New[*rest_model.ServiceDetail](),
+		sessions:           cmap.New[*rest_model.SessionDetail](),
+		serviceEdgeRouters: cmap.New[[]*rest_model.SessionEdgeRouter](),
+		intercepts:         cmap.New[*edge.InterceptV1Config](),
+		EventEmmiter:       events.New(),
 	}
 
 	ctx.CtrlClt = &CtrlClient{